+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0 (2012-XX-XX)
 
+* Added a recursive filter iterator keeping only displayed items
 * Added a filter iterator keeping only current items
 * Added a recursive iterator for the item
 * Fixed building an array of breadcrumbs when a label has only digits